+++ base/markdown/render/terminal/formatting.jl
@@ -51,7 +51,7 @@ end
 # Wrapping
 
 function ansi_length(s)
-    replace(s, r"\e\[[0-9]+m", "") |> length
+    replace(s, r"\e\[[0-9]+m" => "") |> length
 end
 
 words(s) = split(s, " ")