+++ test/libgit2.jl
@@ -2550,7 +2550,7 @@ mktempdir() do dir
             try
                 # In some environments, namely Macs, the hostname "macbook.local" is bound
                 # to the external address while "macbook" is bound to the loopback address.
-                pushfirst!(hostnames, replace(gethostname(), r"\..*$", ""))
+                pushfirst!(hostnames, replace(gethostname(), r"\..*$" => ""))
             end
 
             loopback = ip"127.0.0.1"