+++ test/logging.jl
@@ -241,7 +241,7 @@ end
         s = String(take!(io))
         # Remove the small amount of color, as `Base.print_with_color` can't be
         # simply controlled.
-        s = replace(s, r"^\e\[1m\e\[..m(.*)\e\[39m\e\[22m"m, s"\1")
+        s = replace(s, r"^\e\[1m\e\[..m(.*)\e\[39m\e\[22m"m => s"\1")
         # println(s)
         s
     end